Add accessor for #pos, #pos=
By exposing this method, the CSV gem acted differently and tried sending
the #gets method an extra parameter to limit the bytes returned which it
could not handle. So I also had to handle that parameter which then lead
to an unhandled error because the separator was being passed in as nil
which I was not handling either.